feat: refactor case page to use context
